backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.3
gitpython==3.1.43
PyGithub==2.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.12
